Use Ruby builtin `system` in place of Scriptster
Scriptster's `cmd` doesn't seem to support running commands with an
array as the argument list. And the big benefit we seem to be getting
from it is wrapping all the output lines in log lines. I think it just
makes the output noisier, because Lighthouse already writes
nicely-formatted log lines to stderr, which seem to also confuse
Scriptster and make it show `ERR!` next to every one.

While this was already quoted, so it was unlikely to have issues with
URLs, I think cleaning up the output is worth the change.