Specify version of Lighthouse to use
Since we don't rebuild this container all the time, we already
effectively use a pinned Lighthouse version. And it seems that sometimes
there are bugs in newer Lighthouse versions (we ran into what I think is
a variation of https://github.com/GoogleChrome/lighthouse/issues/6512 on
the current `latest` version). Let's pin this so that we can choose
when to upgrade, rather than upgrading when the container happens to get
rebuilt.