Rename `run.rb` to `bin/lighthouse`
It runs lighthouse, so let's name it appropriately. We're going to start
pulling in other scripts that run more elaborate combinations of things,
and I want the naming to be clear.

This also moves the new `lighthouse` script to a `bin` directory, which
we can use to start gathering all the scripts related to this repo.