Lighthouse: Exclude some audits
These seem to be causing us problems. Lighthouse complains that it
doesn't know what the `full-page-screenshot` audit is, but including
it in the list here does seem to prevent it from being run (dropping
it causes a `FullPageScreenshot` gather to be run). It's been
squirrelly over time, and we don't use the resulting data for the
Lighthouse runs we do with this container, so I'm happy to see it go.