integrate lighthouse and datadog into main script
The main script will iterate through a constant list of urls,
run lighthouse on each url, parse the json output for key metrics to
send to datadog.

That said, we will need a .env file to store the datadog api key and app key,
which also means we need to add .env, along with __pycache__, to .gitignore.