update lighthouse dockerfile
I ran into some issues attempting to run the current image of lighthouse
through the python script. In order to fix this I needed to update the
lighthouse dockerfile.

This updates the lighthouse dockerfile to use Node.js 18-slim instead
of Alpine. This will also install the latest version of lighthouse
which is currently on version 10.3.0.

Lastly this will set a couple of environment variables to configure
Chromium. The `CHROME_PATH` is needed so
that lighthouse will know where to find Chromium. This will also set
the some flags for the chromium browser so we don't need to manually
set them when running lighthouse. The flags will ensure chromium
runs in headless mode and will disable the sandbox which is need when
the container is ran as root. I also needed to
add the flag "--disable-dev-shm-usage" to prevent chromium from
crashing when running in docker.
https://github.com/GoogleChrome/lighthouse/issues/6512#issuecomment-925185412