update lighthouse py class to use docker container
The Lighthouse py class will now execute its command through a lighthouse
docker container instead of directly on the host. This will allow us to
easily run the python code on any host that has docker installed.

That said, this adds some additional checks to ensure that the docker
image is available and to print out the results of the subprocess upon
failure instead of masking it with just an error message.