use json to configure metrics we want to send to datadog
By using a configuration file we can dynamically parse the metrics
from the lighthouse reports and send them to datadog. This will allow
us to easily add/remove metrics without having to modify the code. This
could also be used in the future to modify whether we want to sent
category and/or audit metrics to datadog.