refactor main to audit urls as desktop and mobile
This extracts the core run logic into a separate function that can be
reused with different form factors. This will also tag the results
with the form factor used to audit the URL.

That said, Lighthouse recommends using `--preset=desktop` instead
of `--form-factor=desktop` for the best results. While the
`--form-factor=mobile` is not needed as that is the default, it at least
makes the code more explicit.