feat: retrieve and send scores to DataDog
There are audits that do not contain a numericValue property but do
contain a score property. If want to monitor these audits as well we
will need to retrieve the score property and send it to DataDog.

That said, we will differentiate between the numericValue and score
data by subsetting the metric names with the metric type. For example
instead of just "lighthouse.largest_contentful_paint" we will have
"lighthouse.largest_contentful_paint.score" and
"lighthouse.largest_contentful_paint.value".